requirements.txt
python-telegram-bot==20.3
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0
